+++ abm/abm.js
@@ -217,74 +217,99 @@ function allFilesAreLoaded() {
   // Send text for display in the view
   //postMessage({ command:'text', text:marlin.files.boards.text });
 
+  marlin.watchConfigurations(onConfigFileChanged);
+
+  const version_info = marlin.extractVersionInfo();
+  log("Version Info :", version_info);
+
+  // Post values to the UI filling them in by ID
+  postValue('auth', version_info.auth);
+  postValue('vers', version_info.vers);
+
   const mb = marlin.configValue('MOTHERBOARD');
 
-  if (mb !== undefined) {
-
-    //const sensors = extractTempSensors();
-    //log("Sensors :", sensors);
-
-    const version_info = marlin.extractVersionInfo();
-    log("Version Info :", version_info);
+  if (mb) {
     board_info = marlin.extractBoardInfo(mb);
     log(`Board Info for ${mb} :`, board_info);
-    set_context('has_debug', !!board_info.has_debug);
-    if (board_info.error) {
-      set_context('err.parse', true);
-      postError(board_info.error);
-      return; // abort the whole deal
-    }
-    const machine_info = marlin.getMachineSettings();
-    log("Machine Info :", machine_info);
-    const extruder_info = marlin.getExtruderSettings();
-    log("Extruder Info :", extruder_info);
-    const pindef_info = marlin.getPinDefinitionInfo(mb);
-    log("Pin Defs Info :", pindef_info);
 
-    // If no CUSTOM_MACHINE_NAME was set, get it from the pins file
-    if (!machine_info.name) {
-      let def = marlin.confValue('pindef', 'DEFAULT_MACHINE_NAME');
-      if (!def || def == 'BOARD_INFO_NAME')
-        def = pindef_info.board_name;
-      machine_info.name = def ? def.dequote() : '3D Printer';
-    }
-
-    // Post values to the UI filling them in by ID
-    postValue('auth', version_info.auth);
-    postValue('vers', version_info.vers);
-
-    const d = new Date(version_info.date);
-    postValue('date', d.toLocaleDateString([], { weekday:'long', year:'numeric', month:'short', day:'numeric' }));
-
-    postValue('extruders', extruder_info.extruders);
-    postValue('extruder-desc', extruder_info.description);
-
-    postValue('machine', machine_info.name);
-    postValue('machine-desc', machine_info.description);
-
-    postValue('board', mb.replace('BOARD_', '').replace(/_/g, ' '));
-    postValue('board-desc', board_info.description);
-
-    const pf = board_info.pins_file;
-    const pinsPath = path.join('Marlin', 'src', 'pins', pf);
-    postValue('pins', pf, pinsPath);
-    if (pindef_info.board_name) postValue('pins-desc', pindef_info.board_name);
-
-    postValue('archs', board_info.archs);
-
-    // Fill in the build status in the UI
-    refreshBuildStatus();
+    if (board_info.short)
+      mbn = board_info.short;
+    else
+      mbn = mb.replace('BOARD_', '').replace(/_/g, ' ');
+  }
+  else {
+    mbn = "No MOTHERBOARD";
+    board_info = { error: mbn };
   }
 
-  //marlin.refreshDefineList();
+  if (board_info.error) {
+    set_context('err.parse', true);
+    postError(board_info.error);
+    postValue('date', '');
+    postValue('extruders', '');
+    postValue('extruder-desc', '');
+    postValue('machine', '');
+    postValue('machine-desc', '');
+    postValue('board', mbn);
+    postValue('board-desc', '');
+    postValue('pins', '', '');
+    postValue('pins-desc', '');
+    postValue('archs', ['']);
+    set_context('has_debug', false);
+    board_info.envs = [];
+    refreshBuildStatus();
+    return;
+  }
 
-  marlin.watchConfigurations(onConfigFileChanged);
+  postMessage({ command:'noerror' });
+  const machine_info = marlin.getMachineSettings();
+  log("Machine Info :", machine_info);
+  const extruder_info = marlin.getExtruderSettings();
+  log("Extruder Info :", extruder_info);
+  const pindef_info = marlin.getPinDefinitionInfo(mb);
+  log("Pin Defs Info :", pindef_info);
+  //const sensors = marlin.extractTempSensors();
+  //log("Sensors :", sensors);
+
+  // If no CUSTOM_MACHINE_NAME was set, get it from the pins file
+  if (!machine_info.name) {
+    let def = marlin.confValue('pindef', 'DEFAULT_MACHINE_NAME');
+    if (!def || def == 'BOARD_INFO_NAME')
+      def = pindef_info.board_name;
+    machine_info.name = def ? def.dequote() : '3D Printer';
+  }
+
+  const d = new Date(version_info.date);
+  postValue('date', d.toLocaleDateString([], { weekday:'long', year:'numeric', month:'short', day:'numeric' }));
+
+  postValue('extruders', extruder_info.extruders);
+  postValue('extruder-desc', extruder_info.description);
+
+  postValue('machine', machine_info.name);
+  postValue('machine-desc', machine_info.description);
+
+  postValue('board', mbn);
+  postValue('board-desc', board_info.description);
+
+  const pf = board_info.pins_file;
+  const pinsPath = path.join('Marlin', 'src', 'pins', pf);
+  postValue('pins', pf, pinsPath);
+  if (pindef_info.board_name) postValue('pins-desc', pindef_info.board_name);
+
+  postValue('archs', board_info.archs);
+
+  set_context('has_debug', !!board_info.has_debug);
+
+  // Fill in the build status in the UI
+  refreshBuildStatus();
+
+  //marlin.refreshDefineList(); // TODO: Use the schema object as the one true config data source
   runSelectedAction();
 }
 
 function readFileError(err, msg) {
   log("fs.readFile err: ", err);
-  postMessage({ command:'error', error:msg, data:err });
+  postError(msg, err);
 }
 
 //
@@ -334,7 +359,7 @@ function lastBuild(env) {
     // Find a 'program', .exe, .bin, or .hex file in the folder
     const dirlist = fs.readdirSync(bp);
     const bins = dirlist.filter((n) => {
-      return n.match(/(.+\.(bin|hex|exe|srec)|program|MarlinSimulator)$/i);
+      return n.match(/(.+\.(bin|hex|exe|srec|cbd)|program|MarlinSimulator)$/i);
     });
 
     var tp = bp;
@@ -825,7 +850,9 @@ function runSelectedAction() {
   }
 }
 
-// Look for a script in buildroot/share/PlatformIO and run it.
+// Look for a script in buildroot/share/PlatformIO/scripts and run it.
+// Exit if 'script' is not found or optional 'needs' are not met.
+// Optional 'args' are added to the python command.
 function runPython(script, needs, args) {
   const script_file = path.join(marlin.workspaceRoot, 'buildroot', 'share', 'PlatformIO', 'scripts', script);
   if (!fs.existsSync(script_file)) {